+++ packages/whatsapp_bot_flutter/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 2.0.1
+
+- add wpClient.conn
+- add few group methods
+- fix getChat,getMessages, getContacts
+- minor fixes
+
 ## 2.0.0
 
 - update to whatsapp_bot_platform_interface 1.0.0